+++ udata/__init__.py
@@ -4,5 +4,5 @@
 udata
 '''
 
-__version__ = '5.0.2'
+__version__ = '5.0.3.dev'
 __description__ = 'Open data portal'