+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Add dedicated extras endpoints on resources and datasets [#2779](https://github.com/opendatateam/udata/pull/2779)
+- Enrich catalog with harvest infos [#2789](https://github.com/opendatateam/udata/pull/2789)
 
 ## 5.0.2 (2022-11-29)