+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Current (in progress)
 
+- :warning: Kafka removal [#2783](https://github.com/opendatateam/udata/pull/2783):
+  - Resources events publication now uses webhooks
+  - Search indexation (at runtime and with index command) are now made through HTTP requests
 - Add dedicated extras endpoints on resources and datasets [#2779](https://github.com/opendatateam/udata/pull/2779)
 - Enrich catalog with harvest infos [#2789](https://github.com/opendatateam/udata/pull/2789)
 - Add optionnal license select group custom setting for admin [#2786](https://github.com/opendatateam/udata/pull/2786)