+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add dedicated extras endpoints on resources and datasets [#2779](https://github.com/opendatateam/udata/pull/2779)
 - Enrich catalog with harvest infos [#2789](https://github.com/opendatateam/udata/pull/2789)
 - Add optionnal license select group custom setting for admin [#2786](https://github.com/opendatateam/udata/pull/2786)
+- New feature: Users can now change their email by themselves [#2792](https://github.com/opendatateam/udata/pull/2792)
 
 ## 5.0.2 (2022-11-29)