+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - :warning: Kafka removal [#2783](https://github.com/opendatateam/udata/pull/2783)[#2794](https://github.com/opendatateam/udata/pull/2794):
   - Resources events publication now uses webhooks
-  - Search indexation (at runtime and with index command) are now made through HTTP requests
+  - Search indexation (at runtime and with index command) are now made through HTTP requests.
+  - If you use [udata-search-service](https://pypi.org/project/udata-search-service/), you need to upgrade to the >=2.0.0 version
 - Add dedicated extras endpoints on resources and datasets [#2779](https://github.com/opendatateam/udata/pull/2779)
 - Enrich catalog with harvest infos [#2789](https://github.com/opendatateam/udata/pull/2789)
 - Add optionnal license select group custom setting for admin [#2786](https://github.com/opendatateam/udata/pull/2786)